mudancas de pagina funcinando, faltando apenas estilização para finalizar o projeto
